@@ -17,17 +17,18 @@
   },
   "devDependencies": {
     "babel-eslint": "^3.1.14",
-    "eslint": "^0.22.1",
+    "eslint": "^4.8.0",
     "eslint-plugin-react": "^2.5.0",
     "gulp": "^3.9.0",
     "highlight.js": "~8.8.0",
     "history": "~1.12.5",
+    "prop-types": "^15.6.0",
+    "react": "~0.14.0",
     "react-bootstrap": "~0.27.1",
     "react-component-gulp-tasks": "^0.7.0",
     "react-dom": "~0.14.0",
-    "react-router-bootstrap": "~0.19.2",
     "react-router": "~1.0.0-rc3",
-    "react": "~0.14.0"
+    "react-router-bootstrap": "~0.19.2"
   },
   "peerDependencies": {
     "react": ">=0.14.0",
