@@ -26,6 +26,10 @@ var _classnames2 = _interopRequireDefault(_classnames);
 
 var _reactBootstrap = require('react-bootstrap');
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var Card = (function (_React$Component) {
   _inherits(Card, _React$Component);
 
@@ -53,7 +57,7 @@ var Card = (function (_React$Component) {
 })(_react2['default'].Component);
 
 Card.propTypes = {
-  cardStyle: _react2['default'].PropTypes.oneOf(['default', 'chart', 'summary', 'media', 'file', 'richtext'])
+  cardStyle: _propTypes2['default'].oneOf(['default', 'chart', 'summary', 'media', 'file', 'richtext'])
 };
 
 Card.defaultProps = {
