@@ -26,6 +26,10 @@ var _classnames = require('classnames');
 
 var _classnames2 = _interopRequireDefault(_classnames);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var CardFooter = (function (_React$Component) {
   _inherits(CardFooter, _React$Component);
 
@@ -58,8 +62,8 @@ var CardFooter = (function (_React$Component) {
 })(_react2['default'].Component);
 
 CardFooter.propTypes = {
-  componentClass: _react2['default'].PropTypes.string,
-  footerClassName: _react2['default'].PropTypes.string
+  componentClass: _propTypes2['default'].string,
+  footerClassName: _propTypes2['default'].string
 };
 
 CardFooter.defaultProps = {
