@@ -26,6 +26,10 @@ var _classnames = require('classnames');
 
 var _classnames2 = _interopRequireDefault(_classnames);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var CardHeader = (function (_React$Component) {
   _inherits(CardHeader, _React$Component);
 
@@ -58,8 +62,8 @@ var CardHeader = (function (_React$Component) {
 })(_react2['default'].Component);
 
 CardHeader.propTypes = {
-  componentClass: _react2['default'].PropTypes.string,
-  headerClassName: _react2['default'].PropTypes.string
+  componentClass: _propTypes2['default'].string,
+  headerClassName: _propTypes2['default'].string
 };
 
 CardHeader.defaultProps = {
