@@ -26,6 +26,10 @@ var _classnames = require('classnames');
 
 var _classnames2 = _interopRequireDefault(_classnames);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var CardTitle = (function (_React$Component) {
   _inherits(CardTitle, _React$Component);
 
@@ -58,8 +62,8 @@ var CardTitle = (function (_React$Component) {
 })(_react2['default'].Component);
 
 CardTitle.propTypes = {
-  componentClass: _react2['default'].PropTypes.string,
-  titleClassName: _react2['default'].PropTypes.string
+  componentClass: _propTypes2['default'].string,
+  titleClassName: _propTypes2['default'].string
 };
 
 CardTitle.defaultProps = {
