@@ -28,6 +28,10 @@ var _classnames2 = _interopRequireDefault(_classnames);
 
 var _reactBootstrap = require('react-bootstrap');
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var Cards = (function (_React$Component) {
   _inherits(Cards, _React$Component);
 
@@ -58,7 +62,7 @@ var Cards = (function (_React$Component) {
 })(_react2['default'].Component);
 
 Cards.propTypes = {
-  cardsClassName: _react2['default'].PropTypes.string
+  cardsClassName: _propTypes2['default'].string
 };
 
 Cards.defaultProps = {
