@@ -26,6 +26,10 @@ var _classnames = require('classnames');
 
 var _classnames2 = _interopRequireDefault(_classnames);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var ChartCanvas = (function (_React$Component) {
   _inherits(ChartCanvas, _React$Component);
 
@@ -58,8 +62,8 @@ var ChartCanvas = (function (_React$Component) {
 })(_react2['default'].Component);
 
 ChartCanvas.propTypes = {
-  componentClass: _react2['default'].PropTypes.string,
-  canvasClassName: _react2['default'].PropTypes.string
+  componentClass: _propTypes2['default'].string,
+  canvasClassName: _propTypes2['default'].string
 };
 
 ChartCanvas.defaultProps = {
