@@ -26,6 +26,10 @@ var _classnames = require('classnames');
 
 var _classnames2 = _interopRequireDefault(_classnames);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var Container = (function (_React$Component) {
   _inherits(Container, _React$Component);
 
@@ -58,8 +62,8 @@ var Container = (function (_React$Component) {
 })(_react2['default'].Component);
 
 Container.propTypes = {
-  componentClass: _react2['default'].PropTypes.string,
-  headingClassName: _react2['default'].PropTypes.string
+  componentClass: _propTypes2['default'].string,
+  headingClassName: _propTypes2['default'].string
 };
 
 Container.defaultProps = {
