@@ -28,6 +28,10 @@ var _classnames2 = _interopRequireDefault(_classnames);
 
 var _reactBootstrap = require('react-bootstrap');
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var List = (function (_React$Component) {
   _inherits(List, _React$Component);
 
@@ -58,7 +62,7 @@ var List = (function (_React$Component) {
 })(_react2['default'].Component);
 
 List.propTypes = {
-  listClassName: _react2['default'].PropTypes.string
+  listClassName: _propTypes2['default'].string
 };
 
 List.defaultProps = {
