@@ -26,6 +26,10 @@ var _classnames = require('classnames');
 
 var _classnames2 = _interopRequireDefault(_classnames);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var ListGroupItemText = (function (_React$Component) {
   _inherits(ListGroupItemText, _React$Component);
 
@@ -58,8 +62,8 @@ var ListGroupItemText = (function (_React$Component) {
 })(_react2['default'].Component);
 
 ListGroupItemText.propTypes = {
-  componentClass: _react2['default'].PropTypes.string,
-  textClassName: _react2['default'].PropTypes.string
+  componentClass: _propTypes2['default'].string,
+  textClassName: _propTypes2['default'].string
 };
 
 ListGroupItemText.defaultProps = {
