@@ -26,6 +26,10 @@ var _classnames = require('classnames');
 
 var _classnames2 = _interopRequireDefault(_classnames);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var SummaryItem = (function (_React$Component) {
   _inherits(SummaryItem, _React$Component);
 
@@ -58,8 +62,8 @@ var SummaryItem = (function (_React$Component) {
 })(_react2['default'].Component);
 
 SummaryItem.propTypes = {
-  componentClass: _react2['default'].PropTypes.string,
-  itemClassName: _react2['default'].PropTypes.string
+  componentClass: _propTypes2['default'].string,
+  itemClassName: _propTypes2['default'].string
 };
 
 SummaryItem.defaultProps = {
